+++ modules/lang-mustache/build.gradle
@@ -20,7 +20,7 @@
 esplugin {
   description 'Mustache scripting integration for Elasticsearch'
   classname 'org.elasticsearch.script.mustache.MustachePlugin'
-  hasClientJar = true // For the template query
+  hasClientJar = true // For the template apis and query
 }
 
 dependencies {
@@ -31,9 +31,6 @@ integTest {
   cluster {
     setting 'script.inline', 'true'
     setting 'script.stored', 'true'
-
-
-    File template = new File('src/test/resources/org/elasticsearch/messy/tests/config/scripts/template_1.mustache')
-    extraConfigFile 'scripts/template_1.mustache', template
+    setting 'path.scripts', "${project.buildDir}/resources/test/templates"
   }
 }
