--- modules/lang-mustache/src/test/resources/org/elasticsearch/messy/tests/config/scripts/full-query-template.mustache
@@ -1,6 +1,0 @@
-{
-  "query": {
-    "match": { "{{myField}}" : "{{myValue}}" }
-  },
-  "size" : {{mySize}}
-}
