--- modules/lang-mustache/src/test/resources/org/elasticsearch/messy/tests/config/scripts/storedTemplate.mustache
@@ -1,3 +1,0 @@
-{
-    "match_{{template}}": {}
-}
