+++ modules/lang-mustache/src/test/resources/rest-api-spec/test/lang_mustache/20_render_search_template.yaml
@@ -1,5 +1,5 @@
 ---
-"Indexed Template validate tests":
+"Stored Template validate tests":
 
   - do:
       put_template:
@@ -30,6 +30,23 @@
   - match: { template_output.aggs.my_terms.terms.field: "field1" }
 
 ---
+"File Template validate tests":
+
+  - do:
+      render_search_template:
+        body: { "file": "file_search_template", "params": { "my_value": "foo", "my_field": "field1" } }
+
+  - match: { template_output.query.match.text: "foo" }
+  - match: { template_output.aggs.my_terms.terms.field: "field1" }
+
+  - do:
+      render_search_template:
+        body: { "file": "file_search_template", "params": { "my_value": "bar", "my_field": "my_other_field" } }
+
+  - match: { template_output.query.match.text: "bar" }
+  - match: { template_output.aggs.my_terms.terms.field: "my_other_field" }
+
+---
 "Inline Template validate tests":
 
   - do:
@@ -51,7 +68,7 @@
       render_search_template:
         body: { "inline": { "query": { "match": { "text": "{{{my_value}}" } }, "aggs": { "my_terms": { "terms": { "field": "{{my_field}}" } } } }, "params": { "my_value": "bar", "my_field": "field1" } }
 ---
-"Escaped Indexed Template validate tests":
+"Escaped Stored Template validate tests":
 
   - do:
       put_template:
@@ -102,3 +119,38 @@
       catch: /Improperly.closed.variable.in.query-template/
       render_search_template:
         body: { "inline": "{ \"query\": { \"match\": { \"text\": \"{{{my_value}}\" } }, \"size\": {{my_size}} }", "params": { "my_value": "bar", "my_size": 100 } }
+
+---
+"Indexed Template query tests":
+
+  - do:
+      index:
+        index:  test
+        type:   testtype
+        id:     1
+        body:   { "text": "value1_foo" }
+  - do:
+      index:
+        index:  test
+        type:   testtype
+        id:     2
+        body:   { "text": "value2_foo value3_foo" }
+  - do:
+      indices.refresh: {}
+
+  - do:
+      put_template:
+        id: "1"
+        body: { "template": { "query": { "match" : { "text": "{{my_value}}" } }, "size": "{{my_size}}" } }
+  - match: { acknowledged: true }
+
+  - do:
+      search_template:
+        body: {  "id" : "1", "params" : { "my_value" : "value1_foo", "my_size" : 1 } }
+  - match: { hits.total: 1 }
+
+  - do:
+      catch: /Unable.to.find.on.disk.file.script.\[simple1\].using.lang.\[mustache\]/
+      search_template:
+        body: { "file" : "simple1"}
+
