--- modules/lang-mustache/src/test/resources/rest-api-spec/test/lang_mustache/20_search.yaml
@@ -1,34 +1,0 @@
----
-"Indexed Template query tests":
-
-  - do:
-      index:
-        index:  test
-        type:   testtype
-        id:     1
-        body:   { "text": "value1_foo" }
-  - do:
-      index:
-        index:  test
-        type:   testtype
-        id:     2
-        body:   { "text": "value2_foo value3_foo" }
-  - do:
-      indices.refresh: {}
-
-  - do:
-      put_template:
-        id: "1"
-        body: { "template": { "query": { "match" : { "text": "{{my_value}}" } }, "size": "{{my_size}}" } }
-  - match: { acknowledged: true }
-
-  - do:
-      search_template:
-        body: {  "id" : "1", "params" : { "my_value" : "value1_foo", "my_size" : 1 } }
-  - match: { hits.total: 1 }
-
-  - do:
-      catch: /Unable.to.find.on.disk.file.script.\[simple1\].using.lang.\[mustache\]/
-      search_template:
-        body: { "file" : "simple1"}
-
