--- modules/lang-mustache/src/test/resources/rest-api-spec/test/lang_mustache/40_search_request_template.yaml
@@ -1,38 +1,0 @@
----
-"Template search request":
-
-  - do:
-      index:
-        index:  test
-        type:   testtype
-        id:     1
-        body:   { "text": "value1" }
-  - do:
-      index:
-        index:  test
-        type:   testtype
-        id:     2
-        body:   { "text": "value2" }
-  - do:
-      indices.refresh: {}
-
-  - do:
-      search_template:
-        body: { "inline" : { "query": { "term": { "text": { "value": "{{template}}" } } } }, "params": { "template": "value1" } }
-
-  - match: { hits.total: 1 }
-
-  - do:
-      search_template:
-        body: { "inline" : { "query": { "match_{{template}}": {} } }, "params" : { "template" : "all" } }
-
-  - match: { hits.total: 2 }
-
----
-"Missing template search request":
-
-  - do:
-      catch: missing
-      search_template:
-        body: { "id" : "unknown", "params": { "template": "value1" } }
-
