+++ modules/lang-mustache/src/test/resources/rest-api-spec/test/lang_mustache/40_template_query.yaml
@@ -17,8 +17,32 @@
       indices.refresh: {}
 
   - do:
+      put_template:
+        id: "1"
+        body: { "template": { "match": { "text": "{{my_value}}" } } }
+  - match: { acknowledged: true }
+
+  - do:
       search:
         body: { "query": { "template": { "inline": { "term": { "text": { "value": "{{template}}" } } }, "params": { "template": "value1" } } } }
+
+  - match: { hits.total: 1 }
+
+  - do:
+      search:
+        body: { "query": { "template": { "file": "file_query_template", "params": { "my_value": "value1" } } } }
+
+  - match: { hits.total: 1 }
+
+  - do:
+      search:
+        body: { "query": { "template": { "id": "1", "params": { "my_value": "value1" } } } }
+
+  - match: { hits.total: 1 }
+
+  - do:
+      search:
+        body: { "query": { "template": { "id": "/mustache/1", "params": { "my_value": "value1" } } } }
 
   - match: { hits.total: 1 }
 
@@ -42,7 +66,11 @@
 
   - do:
       search:
+        body: { "query": { "template": { "inline": "{\"match_all\": {}}", "params" : {} } } }
+
+  - match: { hits.total: 2 }
+
+  - do:
+      search:
         body: { "query": { "template": { "inline": "{\"query_string\": { \"query\" : \"{{query}}\" }}", "params" : { "query" : "text:\"value2 value3\"" } } } }
-
-
   - match: { hits.total: 1 }
