+++ modules/lang-mustache/src/test/resources/rest-api-spec/test/lang_mustache/50_multi_search_template.yaml
@@ -124,3 +124,74 @@ setup:
           - inline: '{"query": {"match": {"foo": "{{value}}"} } }'
             params:
               value: "bar"
+
+---
+"Basic multi-search using stored template":
+
+  - do:
+      put_template:
+        id: stored_template_1
+        body: {"template": {"query": {"match": {"{{field}}": "{{value}}" }}}}
+  - match: { acknowledged: true }
+
+  - do:
+      msearch_template:
+        body:
+          - index: index_*
+          - id: stored_template_1
+            params:
+              field: "foo"
+              value: "foo"
+          - index: _all
+          - id: stored_template_1
+            params:
+              field: "foo"
+              value: "bar"
+          - index: index_2
+          - id: stored_template_1
+            params:
+              field: "foo"
+              value: "foo"
+
+  - match:  { responses.0.hits.total:     2  }
+  - match:  { responses.1.hits.total:     1  }
+  - match:  { responses.2.hits.total:     1  }
+
+---
+"Basic multi-search using file template":
+
+  - do:
+      render_search_template:
+        body: { "file": "template_1", "params": { "field": "foo", "value": "bar", "size": 20 } }
+
+  - match: { template_output.query.match.foo.query: "bar" }
+  - match: { template_output.query.match.foo.operator: "or" }
+  - match: { template_output.size: 20 }
+
+  - do:
+      msearch_template:
+        body:
+          - index: index_*
+          - file: template_1
+            params:
+              field: "foo"
+              value: "foo"
+              size: 10
+          - index: _all
+          - file: template_1
+            params:
+              field: "foo"
+              value: "bar"
+              operator: "and"
+              size: 50
+          - index: index_2
+          - file: template_1
+            params:
+              field: "foo"
+              value: "foo"
+              size: 0
+
+  - match:  { responses.0.hits.total:     2  }
+  - match:  { responses.1.hits.total:     1  }
+  - match:  { responses.2.hits.total:     1  }
+
