--- modules/lang-mustache/src/test/resources/rest-api-spec/test/lang_mustache/msearch_template/20_stored_template.yaml
@@ -1,66 +1,0 @@
----
-setup:
-
-  - do:
-      index:
-          index:  index_1
-          type:   test
-          id:     1
-          body:   { foo: bar }
-
-  - do:
-      index:
-          index:  index_1
-          type:   test
-          id:     2
-          body:   { foo: baz }
-
-  - do:
-      index:
-          index:  index_1
-          type:   test
-          id:     3
-          body:   { foo: foo }
-
-  - do:
-      index:
-          index:  index_2
-          type:   test
-          id:     1
-          body:   { foo: foo }
-
-  - do:
-      indices.refresh: {}
-
----
-"Basic multi-search using stored template":
-
-  - do:
-      put_template:
-        id: stored_template_1
-        body: {"template": {"query": {"match": {"{{field}}": "{{value}}" }}}}
-  - match: { acknowledged: true }
-
-  - do:
-      msearch_template:
-        body:
-          - index: index_*
-          - id: stored_template_1
-            params:
-              field: "foo"
-              value: "foo"
-          - index: _all
-          - id: stored_template_1
-            params:
-              field: "foo"
-              value: "bar"
-          - index: index_2
-          - id: stored_template_1
-            params:
-              field: "foo"
-              value: "foo"
-
-  - match:  { responses.0.hits.total:     2  }
-  - match:  { responses.1.hits.total:     1  }
-  - match:  { responses.2.hits.total:     1  }
-
