--- modules/lang-mustache/src/test/resources/rest-api-spec/test/lang_mustache/msearch_template/30_file_template.yaml
@@ -1,72 +1,0 @@
----
-setup:
-
-  - do:
-      index:
-          index:  index_1
-          type:   test
-          id:     1
-          body:   { foo: bar }
-
-  - do:
-      index:
-          index:  index_1
-          type:   test
-          id:     2
-          body:   { foo: baz }
-
-  - do:
-      index:
-          index:  index_1
-          type:   test
-          id:     3
-          body:   { foo: foo }
-
-  - do:
-      index:
-          index:  index_2
-          type:   test
-          id:     1
-          body:   { foo: foo }
-
-  - do:
-      indices.refresh: {}
-
----
-"Basic multi-search using file template":
-
-  - do:
-      render_search_template:
-        body: { "file": "template_1", "params": { "field": "foo", "value": "bar", "size": 20 } }
-
-  - match: { template_output.query.match.foo.query: "bar" }
-  - match: { template_output.query.match.foo.operator: "or" }
-  - match: { template_output.size: 20 }
-
-  - do:
-      msearch_template:
-        body:
-          - index: index_*
-          - file: template_1
-            params:
-              field: "foo"
-              value: "foo"
-              size: 10
-          - index: _all
-          - file: template_1
-            params:
-              field: "foo"
-              value: "bar"
-              operator: "and"
-              size: 50
-          - index: index_2
-          - file: template_1
-            params:
-              field: "foo"
-              value: "foo"
-              size: 0
-
-  - match:  { responses.0.hits.total:     2  }
-  - match:  { responses.1.hits.total:     1  }
-  - match:  { responses.2.hits.total:     1  }
-
