+++ modules/lang-mustache/src/test/resources/templates/file_query_template.mustache
@@ -1,0 +1,5 @@
+{
+    "match": {
+        "text": "{{my_value}}"
+    }
+}
