+++ modules/lang-mustache/src/test/resources/templates/file_search_template.mustache
@@ -1,0 +1,12 @@
+{
+    "query": {
+        "match": {
+            "text": "{{my_value}}"
+        }
+    },
+    "aggs": {
+        "my_terms": {
+            "terms": { "field": "{{my_field}}" }
+        }
+    }
+}
