+++ modules/lang-mustache/src/test/resources/templates/template_1.mustache
